fix: check existing group names when naming new group
Previous behaviour was naming a new group `Group n`, where n is the
total number of groups + 1. New behaviour is to start with n = 1 and
check if the groups list contains `Group n`. If it doesn't, the name
is available, and if it does, add one to n and check again.